ErrorResponse implement Error interface
ErrorResponse add two methods for biz  error check

add HasBizError() bool

add Error() string
